+++ src/RadixDlt.NetworkGateway.GatewayApi/Validators/StateEntityNonFungiblesPageRequestValidator.cs
@@ -75,12 +75,13 @@ internal class StateEntityNonFungiblesPageRequestValidator : AbstractValidator<G
         IOptionsSnapshot<EndpointOptions> endpointOptionsSnapshot,
         LedgerStateSelectorValidator ledgerStateSelectorValidator,
         PaginableRequestValidator paginableRequestValidator,
+        RadixAddressValidator radixAddressValidator,
         StateEntityNonFungiblesPageRequestOptInsValidator nonFungiblesPageRequestOptInsValidator
             )
     {
         RuleFor(x => x.Address)
             .NotEmpty()
-            .RadixAddress();
+            .SetValidator(radixAddressValidator);
 
         RuleFor(x => x.AggregationLevel)
             .IsInEnum();